editor: When in Editor Mode Open in Same Window.
This means when an exiting editor sessions is open (in non-project
mode), Edi will open additional files within the same Edi
session. Editor mode, is as a "ordinary" text editor.

If there is a problem, a timeout will occur and Edi will open as
normal. There shouldn't be a problem :/